# packages/api-contracts/openapi/bff-v1.yaml
openapi: 3.0.3
info:
  title: BetoGo BFF API
  version: 1.1.0
  description: |
    C 端 BFF v0.1 — 29 endpoints per docs/product/BFF-CLIENT-API.md
servers:
  - url: http://localhost:3000/api/v1
    description: Local Docker / dev
tags:
  - name: auth
  - name: user
  - name: wallet
  - name: deposits
  - name: withdrawals
  - name: ledger
  - name: kyc
  - name: promotions
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    ApiResponse:
      type: object
      properties:
        code: { type: integer, example: 0 }
        message: { type: string, example: ok }
        data: {}
        traceId: { type: string }
    AuthSession:
      type: object
      properties:
        token: { type: string }
        expiresIn: { type: integer }
        isNewUser: { type: boolean }
        trialRedPacketEligible: { type: boolean }
        user:
          type: object
          properties:
            id: { type: string }
            telegramUserId: { type: integer }
            displayName: { type: string }
            avatarUrl: { type: string }
            inviteCode: { type: string }
paths:
  /auth/telegram:
    post:
      tags: [auth]
      summary: Telegram initData login
      parameters:
        - in: header
          name: X-Telegram-Init-Data
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                initData: { type: string }
                start_param: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - properties:
                      data: { $ref: '#/components/schemas/AuthSession' }
  /auth/session:
    get:
      tags: [auth]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Session validity
  /auth/refresh:
    post:
      tags: [auth]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: New token
  /auth/logout:
    post:
      tags: [auth]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Logged out
  /user/me:
    get:
      tags: [user]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: User profile
    patch:
      tags: [user]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Updated profile
  /user/status:
    get:
      tags: [user]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Account status
  /user/language:
    patch:
      tags: [user]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Locale updated
  /wallet/balances:
    get:
      tags: [wallet]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Balance list (PHP cents)
  /wallet/summary:
    get:
      tags: [wallet]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Wallet summary
  /wallet/turnover:
    get:
      tags: [wallet]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Turnover progress
  /deposits:
    post:
      tags: [deposits]
      security: [{ bearerAuth: [] }]
      summary: Create deposit (channelId=tg_wallet)
      responses:
        '200':
          description: Order created
    get:
      tags: [deposits]
      security: [{ bearerAuth: [] }]
      summary: Deposit history
      responses:
        '200':
          description: Paginated list
  /deposits/{orderId}:
    get:
      tags: [deposits]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: orderId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Deposit status
  /withdrawals/eligibility:
    get:
      tags: [withdrawals]
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: query, name: currency, schema: { type: string } }
        - { in: query, name: channelId, schema: { type: string, enum: [tg_wallet] } }
        - { in: query, name: amount, schema: { type: integer } }
      responses:
        '200':
          description: Eligibility check
  /withdrawals:
    post:
      tags: [withdrawals]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Withdrawal created
    get:
      tags: [withdrawals]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Withdrawal history
  /withdrawals/{orderId}:
    get:
      tags: [withdrawals]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: orderId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Withdrawal detail
  /ledger:
    get:
      tags: [ledger]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Ledger list
  /ledger/{id}:
    get:
      tags: [ledger]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Ledger entry
  /kyc/status:
    get:
      tags: [kyc]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: KYC status
  /kyc/documents/upload-url:
    post:
      tags: [kyc]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Upload credentials
  /kyc/submissions:
    post:
      tags: [kyc]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: KYC submitted
  /kyc/submissions/latest:
    get:
      tags: [kyc]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Latest submission
  /promotions:
    get:
      tags: [promotions]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Promotion list
  /promotions/{promoId}:
    get:
      tags: [promotions]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: promoId
          schema: { type: string, enum: [trial, referral, firstdep] }
          required: true
      responses:
        '200':
          description: Promotion detail
    post:
      tags: [promotions]
      security: [{ bearerAuth: [] }]
      summary: Claim promotion reward
      parameters:
        - in: path
          name: promoId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Claimed
  /promotions/trial-play:
    get:
      tags: [promotions]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Trial progress
  /promotions/trial-play/claim:
    post:
      tags: [promotions]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Trial red packet claimed
  /promotions/referral:
    get:
      tags: [promotions]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Referral summary
  /promotions/referral/link:
    get:
      tags: [promotions]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Invite deep link
  /promotions/referral/records:
    get:
      tags: [promotions]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Referral records
  /promotions/red-packets:
    get:
      tags: [promotions]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Red packet history